v1.6.26a

+++ deepauth/views.py
@@ -1,5 +1,5 @@
 from datetime import timedelta
-
+from deeputils.schemas import RefinedViewSet
 from captcha.helpers import captcha_image_url
 from captcha.models import CaptchaStore
 from django.contrib.auth import authenticate
@@ -10,7 +10,6 @@ from rest_framework.authtoken.models import Token
 from rest_framework.exceptions import ParseError, NotAcceptable
 from rest_framework.permissions import AllowAny, IsAuthenticated, IsAdminUser
 from rest_framework.response import Response
-from rest_framework.views import APIView
 
 from deepauth.serializers import *
 from deepauth.utils.mailbox import send_mail
@@ -18,27 +17,34 @@ from deepauth.utils.password import change_password, auth_password
 from deepauth.utils.token import TOKEN_LIFETIME, ExpiringTokenAuthentication
 
 
-class RegisterView(APIView):
+class RegisterView(RefinedViewSet):
     """
-    post:
-    **注册用户**
-    - <span class='badge'>R</span> `password` 密码，建议为 MD5 哈希结果
-    - <span class='badge'>R</span> `first_name` 用户称呼（名），不能超过 30 个字符
-    - `username` 用户名，不能超过 150 个字符，如不提供则会依照当前时间生成一个
-    - `last_name` 用户称呼（姓），不能超过 30 个字符
-    - `email` 邮箱
-    - `tel` 手机号码
-    - `country` 国家
-    - `invitation_code` 邀请码
-    - `captcha_key` 验证码的哈希值，建议隐藏，失效时间为 5 分钟
-    - `captcha_value` 验证码的答案
+        create:
+        **注册用户**
+    
+        post:
+        **注册用户**
+        - <span class='badge'>R</span> `password` 密码，建议为 MD5 哈希结果
+        - <span class='badge'>R</span> `first_name` 用户称呼（名），不能超过 30 个字符
+        - `username` 用户名，不能超过 150 个字符，如不提供则会依照当前时间生成一个
+        - `last_name` 用户称呼（姓），不能超过 30 个字符
+        - `email` 邮箱
+        - `tel` 手机号码
+        - `country` 国家
+        - `invitation_code` 邀请码
+        - `captcha_key` 验证码的哈希值，建议隐藏，失效时间为 5 分钟
+        - `captcha_value` 验证码的答案
     """
+
     authentication_classes = ()
     permission_classes = (AllowAny,)
-    serializer_class = RegisterViewSerializer
 
-    def post(self, request):
-        pp = self.serializer_class(data=request.data)
+    serializer_classes = {
+        'create': RegisterViewSerializer,
+    }
+
+    def create(self, request):
+        pp = self.get_serializer(data=request.data)
         if pp.is_valid():
             first_name = pp.validated_data['first_name']
             last_name = pp.validated_data['last_name']
@@ -65,7 +71,7 @@ class RegisterView(APIView):
                 invitation_code.save()
             # 赠送邀请码
             for i in range(INVITATION_LIMIT):
-            # for i in range(getattr(settings, 'DEEPAUTH_INVITATION_ONLY', 10)):
+                # for i in range(getattr(settings, 'DEEPAUTH_INVITATION_ONLY', 10)):
                 invitation_code = InvitationCode(account=account)
                 invitation_code.save()
             return Response({'id': account.id, 'username': account.username}, status=status.HTTP_201_CREATED)
@@ -73,21 +79,28 @@ class RegisterView(APIView):
             raise ParseError(pp.errors)
 
 
-class LoginView(APIView):
+class LoginView(RefinedViewSet):
     """
-    get:
-    **登录**
-    - <span class='badge'>R</span> `certification` 用户名或邮箱或手机号
-    - <span class='badge'>R</span> `password` 密码，建议为 MD5 哈希结果
-    - `captcha_key` 验证码的哈希值，建议隐藏，失效时间为 5 分钟
-    - `captcha_value` 验证码的答案
+        list:
+        **登录**
+
+        get:
+        **登录**
+        - <span class='badge'>R</span> `certification` 用户名或邮箱或手机号
+        - <span class='badge'>R</span> `password` 密码，建议为 MD5 哈希结果
+        - `captcha_key` 验证码的哈希值，建议隐藏，失效时间为 5 分钟
+        - `captcha_value` 验证码的答案
     """
+
     authentication_classes = ()
     permission_classes = (AllowAny,)
-    serializer_class = LoginViewSerializer
 
-    def get(self, request):
-        pp = self.serializer_class(data=request.GET)
+    serializer_classes = {
+        'list': LoginViewSerializer,
+    }
+
+    def list(self, request):
+        pp = self.get_serializer(data=request.GET)
         if pp.is_valid():
             certification = pp.validated_data['certification']
             password = pp.validated_data['password']
@@ -107,41 +120,63 @@ class LoginView(APIView):
             raise ParseError(pp.errors)
 
 
-class LogoutView(APIView):
+class LogoutView(RefinedViewSet):
     """
-    post:
-    <span class='badge'><i class='fa fa-lock'></i></span> **登出**
+        create:
+        **登出**
+
+        post:
+        <span class='badge'><i class='fa fa-lock'></i></span> **登出**
     """
+
     authentication_classes = (ExpiringTokenAuthentication, BasicAuthentication)
     permission_classes = (IsAuthenticated,)
 
-    def post(self, request):
-        request.user.auth_token.delete()
-        return Response()
+    serializer_classes = {
+        'create': LogoutViewSerializer,
+    }
+
+    def create(self, request):
+        pp = self.get_serializer(data=request.data)
+        if pp.is_valid():
+            request.user.auth_token.delete()
+            return Response()
+
+        else:
+            raise ParseError(pp.errors)
 
 
-class PasswordView(APIView):
+class PasswordView(RefinedViewSet):
     """
-    get:
-    <span class='badge'><i class='fa fa-lock'></i></span> **获取用户密码修改历史**
+        get:
+        <span class='badge'><i class='fa fa-lock'></i></span> **获取用户密码修改历史**
 
-    put:
-    <span class='badge'><i class='fa fa-lock'></i></span> **修改密码**
+        put:
+        <span class='badge'><i class='fa fa-lock'></i></span> **修改密码**
 
-    - <span class='badge'>R</span> `password_old` 当前密码
-    - <span class='badge'>R</span> `password_new` 新密码
-    - <span class='badge'>R</span> `password_confirm` 重复新密码
+        - <span class='badge'>R</span> `password_old` 当前密码
+        - <span class='badge'>R</span> `password_new` 新密码
+        - <span class='badge'>R</span> `password_confirm` 重复新密码
     """
+
     authentication_classes = (ExpiringTokenAuthentication, BasicAuthentication)
     permission_classes = (IsAuthenticated,)
-    serializer_class = PasswordViewSerializer
 
-    def get(self, request):
-        account = Account.objects.get(pk=request.user.pk)
-        return Response(PasswordLogSerializer(PasswordLog.objects.filter(account=account)[:10], many=True).data)
+    serializer_classes = {
+        'list': PasswordGetViewSerializer,
+        'update': PasswordPutViewSerializer,
+    }
 
-    def put(self, request):
-        pp = self.serializer_class(data=request.data)
+    def list(self, request):
+        pp = self.get_serializer(data=request.GET)
+        if pp.is_valid():
+            account = Account.objects.get(pk=request.user.pk)
+            return Response(PasswordLogSerializer(PasswordLog.objects.filter(account=account)[:10], many=True).data)
+        else:
+            raise ParseError(pp.errors)
+
+    def update(self, request):
+        pp = self.get_serializer(data=request.data)
         if pp.is_valid():
             username = request.user.username
             password_old = pp.validated_data['password_old']
@@ -158,33 +193,42 @@ class PasswordView(APIView):
             raise ParseError(pp.errors)
 
 
-class DetailView(APIView):
+class DetailView(RefinedViewSet):
     """
-    get:
-    <span class='badge'><i class='fa fa-lock'></i></span> **获取用户信息**
+        get:
+        <span class='badge'><i class='fa fa-lock'></i></span> **获取用户信息**
 
-    put:
-    <span class='badge'><i class='fa fa-lock'></i></span> **修改用户信息**
+        put:
+        <span class='badge'><i class='fa fa-lock'></i></span> **修改用户信息**
 
-    - <span class='badge'>R</span> `field` 修改键值，逗号分隔：`unique_auth` 是否仅限单一客户端登录、
-    `email` 邮箱、`last_name` 用户称呼（姓）、`first_name` 用户称呼（名）、`avatar` 用户头像、
-    `country` 国家、`tel` 手机号码
-    - <span class='badge'>R</span> `value` 修改内容，逗号分隔，必须与 `field` 长度相同
+        - <span class='badge'>R</span> `field` 修改键值，逗号分隔：`unique_auth` 是否仅限单一客户端登录、
+        `email` 邮箱、`last_name` 用户称呼（姓）、`first_name` 用户称呼（名）、`avatar` 用户头像、
+        `country` 国家、`tel` 手机号码
+        - <span class='badge'>R</span> `value` 修改内容，逗号分隔，必须与 `field` 长度相同
     """
+
     authentication_classes = (ExpiringTokenAuthentication, BasicAuthentication)
     permission_classes = (IsAuthenticated,)
-    serializer_class = DetailViewSerializer
 
-    def get(self, request):
-        account = Account.objects.get(pk=request.user.pk)
-        resp = AccountSerializer(account).data
-        resp['access_log'] = AccessLogSerializer(AccessLog.objects.filter(account=account)[:10], many=True).data
-        resp['invitation_code'] = InvitationCodeSerializer(InvitationCode.objects.filter(account=account, user=None), many=True).data
-        return Response(resp)
+    serializer_classes = {
+        'list': DetailGetViewSerializer,
+        'update': DetailPutViewSerializer,
+    }
 
-    def put(self, request):
+    def list(self, request):
+        pp = self.get_serializer(data=request.GET)
+        if pp.is_valid():
+            account = Account.objects.get(pk=request.user.pk)
+            resp = AccountSerializer(account).data
+            resp['access_log'] = AccessLogSerializer(AccessLog.objects.filter(account=account)[:10], many=True).data
+            resp['invitation_code'] = InvitationCodeSerializer(InvitationCode.objects.filter(account=account, user=None), many=True).data
+            return Response(resp)
+        else:
+            raise ParseError(pp.errors)
+
+    def update(self, request):
         account = Account.objects.get(pk=request.user.pk)
-        pp = self.serializer_class(data=request.data)
+        pp = self.get_serializer(data=request.data)
         if pp.is_valid():
             if 'email' in pp.validated_data:
                 email = pp.validated_data['email']
@@ -202,29 +246,50 @@ class DetailView(APIView):
             raise ParseError(pp.errors)
 
 
-class AdminAccountView(APIView):
+class AdminAccountView(RefinedViewSet):
     """
     get:
-    <span class='badge'><i class='fa fa-lock'></i></span> <span class='badge'><i class='fa fa-cog'></i></span> **获取全部用户信息**
+        <span class='badge'><i class='fa fa-lock'></i></span> <span class='badge'><i class='fa fa-cog'></i></span> **获取全部用户信息**
 
-    put:
-    <span class='badge'><i class='fa fa-lock'></i></span> <span class='badge'><i class='fa fa-cog'></i></span> **修改用户信息**
+        put:
+        <span class='badge'><i class='fa fa-lock'></i></span> <span class='badge'><i class='fa fa-cog'></i></span> **修改用户信息**
 
-    - <span class='badge'>R</span> `id` 用户 ID
-    - <span class='badge'>R</span> `field` 修改键值，逗号分隔：`unique_auth` 是否仅限单一客户端登录、
-    `email` 邮箱、`last_name` 用户称呼（姓）、`first_name` 用户称呼（名）、`avatar` 用户头像、`country` 国家、
-    `tel` 手机号码、`password` 密码、`is_active` 是否活跃
-    - <span class='badge'>R</span> `value` 修改内容，逗号分隔，必须与 `field` 长度相同
+        - <span class='badge'>R</span> `id` 用户 ID
+        - <span class='badge'>R</span> `field` 修改键值，逗号分隔：`unique_auth` 是否仅限单一客户端登录、
+        `email` 邮箱、`last_name` 用户称呼（姓）、`first_name` 用户称呼（名）、`avatar` 用户头像、`country` 国家、
+        `tel` 手机号码、`password` 密码、`is_active` 是否活跃
+        - <span class='badge'>R</span> `value` 修改内容，逗号分隔，必须与 `field` 长度相同
+
+        get:
+        <span class='badge'><i class='fa fa-lock'></i></span> <span class='badge'><i class='fa fa-cog'></i></span> **获取全部用户信息**
+
+        put:
+        <span class='badge'><i class='fa fa-lock'></i></span> <span class='badge'><i class='fa fa-cog'></i></span> **修改用户信息**
+
+        - <span class='badge'>R</span> `id` 用户 ID
+        - <span class='badge'>R</span> `field` 修改键值，逗号分隔：`unique_auth` 是否仅限单一客户端登录、
+        `email` 邮箱、`last_name` 用户称呼（姓）、`first_name` 用户称呼（名）、`avatar` 用户头像、`country` 国家、
+        `tel` 手机号码、`password` 密码、`is_active` 是否活跃
+        - <span class='badge'>R</span> `value` 修改内容，逗号分隔，必须与 `field` 长度相同
     """
+
     authentication_classes = (ExpiringTokenAuthentication, BasicAuthentication)
     permission_classes = (IsAdminUser,)
-    serializer_class = AdminAccountViewSerializer
 
-    def get(self, request):
-        return Response(AccountSerializer(Account.objects.all(), many=True).data)
+    serializer_classes = {
+        'list': AdminAccountGetViewSerializer,
+        'update': AdminAccountPutViewSerializer,
+    }
 
-    def put(self, request):
-        pp = self.serializer_class(data=request.data)
+    def list(self, request):
+        pp = self.get_serializer(data=request.GET)
+        if pp.is_valid():
+            return Response(AccountSerializer(Account.objects.all(), many=True).data)
+        else:
+            raise ParseError(pp.errors)
+
+    def update(self, request):
+        pp = self.get_serializer(data=request.data)
         if pp.is_valid():
             uid = pp.validated_data['id']
             account = Account.objects.get(pk=uid)
@@ -249,7 +314,7 @@ class AdminAccountView(APIView):
             raise ParseError(pp.errors)
 
 
-# class AdminAccountTreeView(APIView):
+# class AdminAccountTreeView(RefinedViewSet):
 #     """
 #     get:
 #     - <span class='badge'>R</span> `id` 用户 ID
@@ -274,18 +339,23 @@ class AdminAccountView(APIView):
 #             raise ParseError(pp.errors)
 
 
-class ActivateEmailView(APIView):
+class ActivateEmailView(RefinedViewSet):
     """
-    get:
-    - <span class='badge'>R</span> `id` 用户 id ,
-    - <span class='badge'>R</span> `prefix` url 地址
+        get:
+        - <span class='badge'>R</span> `id` 用户 id ,
+        - <span class='badge'>R</span> `prefix` url 地址
     """
+
     authentication_classes = ()
     permission_classes = (AllowAny,)
     serializer_class = ActivateEmailViewSerializer
 
-    def get(self, request):
-        pp = self.serializer_class(data=request.GET)
+    serializer_classes = {
+        'list': ActivateEmailViewSerializer,
+    }
+
+    def list(self, request):
+        pp = self.get_serializer(data=request.GET)
         if pp.is_valid():
             uid = pp.validated_data['id']
             prefix = pp.validated_data['prefix']
@@ -310,18 +380,22 @@ class ActivateEmailView(APIView):
             raise ParseError(pp.errors)
 
 
-class ValidateEmailView(APIView):
+class ValidateEmailView(RefinedViewSet):
     """
-    get:
-    - <span class='badge'>R</span> `id` 用户 ID ,
-    - <span class='badge'>R</span> `code` 用户的激活码
+        get:
+        - <span class='badge'>R</span> `id` 用户 ID ,
+        - <span class='badge'>R</span> `code` 用户的激活码
     """
+
     authentication_classes = ()
     permission_classes = (AllowAny,)
-    serializer_class = ValidateEmailViewSerializer
 
-    def get(self, request):
-        pp = self.serializer_class(data=request.GET)
+    serializer_classes = {
+        'list': ValidateEmailViewSerializer,
+    }
+
+    def list(self, request):
+        pp = self.get_serializer(data=request.GET)
         if pp.is_valid():
             uid = pp.validated_data['id']
             code = pp.validated_data['code']
@@ -355,19 +429,31 @@ class ValidateEmailView(APIView):
             raise ParseError(pp.errors)
 
 
-class CaptchaView(APIView):
+class CaptchaView(RefinedViewSet):
     """
+        list:
+        **获取验证码**
+
         get:
         <span class='badge'><i class='fa fa-lock'></i></span> **获取验证码**
     """
+
     authentication_classes = ()
     permission_classes = (AllowAny,)
 
-    def get(self, request):
-        CaptchaStore.remove_expired()  # 删除失效的验证码，过期时间为 5 分钟
-        captcha_key = CaptchaStore.pick()
-        to_json_response = {
-            'key': captcha_key,
-            'url': captcha_image_url(captcha_key),
-        }
-        return Response(to_json_response)
+    serializer_classes = {
+        'list': CaptchaGetViewSerializer,
+    }
+
+    def list(self, request):
+        pp = self.get_serializer(data=request.GET)
+        if pp.is_valid():
+            CaptchaStore.remove_expired()  # 删除失效的验证码，过期时间为 5 分钟
+            captcha_key = CaptchaStore.pick()
+            to_json_response = {
+                'key': captcha_key,
+                'url': captcha_image_url(captcha_key),
+            }
+            return Response(to_json_response)
+        else:
+            raise ParseError(pp.errors)
